requirements.txt
pytest
selenium
pytest-html
pytest-xdist
allure-pytest
openpyxl
matplotlib
pandas
numpy
Faker
pytest-check

